refactor: split github_release.rs god-file into modules
