+++ src/core/release/executor/github_release/gh_cli.rs
@@ -1,0 +1,156 @@
+//! `gh` CLI probes, environment, command construction, and path/quote helpers.
+
+use crate::core::component::GithubConfig;
+use crate::core::deploy::release_download::GitHubRepo;
+use crate::core::release::types::ReleaseState;
+
+pub(crate) fn gh_is_available() -> bool {
+    crate::core::git::gh_probe_succeeds(&["--version"])
+}
+
+pub(crate) fn gh_is_authenticated(github: &GitHubRepo, config: &GithubConfig) -> bool {
+    gh_probe_succeeds(
+        github,
+        config,
+        &["auth", "status", "--hostname", &github.host],
+    )
+}
+
+pub(crate) fn gh_release_exists(
+    github: &GitHubRepo,
+    config: &GithubConfig,
+    tag: &str,
+    repo_flag: &str,
+) -> bool {
+    gh_probe_succeeds(github, config, &["release", "view", tag, "-R", repo_flag])
+}
+
+pub(crate) fn github_release_artifact_paths(state: &ReleaseState) -> Vec<String> {
+    state
+        .artifacts
+        .iter()
+        .filter_map(|artifact| {
+            artifact
+                .durable_path
+                .as_deref()
+                .filter(|path| path_is_file(path))
+                .or(Some(artifact.path.as_str()))
+                .filter(|path| path_is_file(path))
+                .map(str::to_string)
+        })
+        .collect()
+}
+
+fn path_is_file(path: &str) -> bool {
+    std::fs::metadata(path)
+        .map(|metadata| metadata.is_file())
+        .unwrap_or(false)
+}
+
+pub(super) fn gh_env_prefix(env: &[(String, String)]) -> String {
+    let parts = env
+        .iter()
+        .filter(|(key, value)| !key.is_empty() && !value.is_empty())
+        .map(|(key, value)| format!("{}={}", key, shell_quote(value)))
+        .collect::<Vec<_>>();
+    if parts.is_empty() {
+        String::new()
+    } else {
+        format!("{} ", parts.join(" "))
+    }
+}
+
+pub(super) fn gh_env_hint(github: &GitHubRepo, env: &[(String, String)]) -> Option<String> {
+    if github.host == "github.com" && env.is_empty() {
+        return None;
+    }
+
+    let mut hints = Vec::new();
+    let proxy_keys = env
+        .iter()
+        .filter(|(key, value)| is_proxy_env_key(key) && !value.is_empty())
+        .map(|(key, _)| key.as_str())
+        .collect::<Vec<_>>();
+    if github.host != "github.com" {
+        hints.push(format!(
+            "GitHub Enterprise host detected: repair commands include GH_HOST={}",
+            github.host
+        ));
+    }
+    if !proxy_keys.is_empty() {
+        hints.push(format!(
+            "Proxy environment is included in repair commands: {}.",
+            proxy_keys.join(", ")
+        ));
+    } else if github.host != "github.com" {
+        hints.push(
+            "If this Enterprise host requires a proxy, prefix the commands with the needed HTTPS_PROXY/HTTP_PROXY/ALL_PROXY value.".to_string(),
+        );
+    }
+
+    Some(hints.join(" "))
+}
+
+fn is_proxy_env_key(key: &str) -> bool {
+    matches!(
+        key,
+        "HTTPS_PROXY" | "https_proxy" | "HTTP_PROXY" | "http_proxy" | "ALL_PROXY" | "all_proxy"
+    )
+}
+
+pub(super) fn safe_filename(value: &str) -> String {
+    value
+        .chars()
+        .map(|c| {
+            if c.is_ascii_alphanumeric() || matches!(c, '.' | '-' | '_') {
+                c
+            } else {
+                '-'
+            }
+        })
+        .collect()
+}
+
+pub(super) fn shell_quote(value: &str) -> String {
+    if value
+        .chars()
+        .all(|c| c.is_ascii_alphanumeric() || matches!(c, '/' | '.' | '_' | '-' | ':' | '=' | '@'))
+    {
+        return value.to_string();
+    }
+
+    format!("'{}'", value.replace('\'', "'\\''"))
+}
+
+fn gh_probe_succeeds(github: &GitHubRepo, config: &GithubConfig, args: &[&str]) -> bool {
+    command_probe_succeeds(gh_command(github, config, args))
+}
+
+/// Run a prepared command swallowing stdout/stderr and report whether it exited
+/// successfully. Centralizes the probe-style `null stdio + status + success`
+/// pattern so probe call sites do not each reimplement it.
+fn command_probe_succeeds(mut command: std::process::Command) -> bool {
+    command
+        .stdout(std::process::Stdio::null())
+        .stderr(std::process::Stdio::null())
+        .status()
+        .map(|s| s.success())
+        .unwrap_or(false)
+}
+
+pub(super) fn gh_command(
+    github: &GitHubRepo,
+    config: &GithubConfig,
+    args: &[&str],
+) -> std::process::Command {
+    let mut command = std::process::Command::new("gh");
+    command.args(args);
+    for (key, value) in github_cli_env(github, config) {
+        command.env(key, value);
+    }
+    command
+}
+
+pub(crate) fn github_cli_env(github: &GitHubRepo, config: &GithubConfig) -> Vec<(String, String)> {
+    crate::core::git::github_cli_env(&github.host, config)
+}
